backend/requirements.txt
fastapi==0.115.6

uvicorn[standard]==0.32.1

pydantic==2.9.2

httpx==0.28.1

openai==1.59.9

pypdf==5.1.0

python-dotenv==1.0.1


